Split into entity into model and entity, fixed issue with validation username by adding tag
